requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery[redis]==5.3.6
django-redis==5.4.0
redis==5.0.1
httpx==0.27.0
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
gym==0.26.2
tensorflow==2.15.0
gunicorn==21.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
